# api/swagger.yaml
openapi: "3.0.3"
info:
  title: Mailbadger API
  version: "1.0"
  contact:
    name: Mailbadger support team
    email: support@mailbadger.io
  description: |
    The Mailbadger [REST](https://en.wikipedia.org/wiki/Representational_state_transfer) API.
  license:
    name: “Commons Clause” License Condition v1.0
    url: https://github.com/mailbadger/app/blob/master/LICENSE
servers:
  - url: https://app.mailbadger.io/api
...
paths:
...
  /subscribers/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      tags:
        - subscribers
      operationId: getSubscriber
      summary: Get subscriber by ID
      description: Returns a single subscriber object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                message: Subscriber not found.
        default:
          $ref: "#/components/responses/UnexpectedError"
    put:
      tags:
        - subscribers
      operationId: updateSubscriber
      summary: Update an existing subscriber
      description: Update an existing subscriber.
      requestBody:
        $ref: "#/components/requestBodies/UpdateSubscriberParams"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscriber"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                message: Subscriber not found.
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Message"
                  - $ref: "#/components/schemas/ValidationErrors"
              examples:
                validationErrors:
                  summary: Invalid parameters
                  value:
                    message: Invalid parameters, please try again
                    errors:
                      name: The subscriber name is too long.
                message:
                  summary: Bad request
                  value:
                    message: Subscriber could not be created, please try again.
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                message: Subscriber with that email already exist.
        default:
          $ref: "#/components/responses/UnexpectedError"
    delete:
      tags:
        - subscribers
      operationId: deleteSubscriber
      summary: Delete a subscriber
      description: Delete a subscriber.
      responses:
        "204":
          description: The subscriber was deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                message: Invalid ID supplied.
        default:
          $ref: "#/components/responses/UnexpectedError"
...
security:
  - api_key: []
components:
  requestBodies:
...
    UpdateSubscriberParams:
      description: Parameters for the update subscriber form
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
                example: John Doe
                description: The subscriber's name.
                maxLength: 191
              segments:
                type: array
                description: Groups to place the subscriber in.
                items:
                  type: integer
              metadata:
                description: |
                  Metadata (key=value pairs) for the subscriber. These values can be used in the campaign templates to 
                  personalize the e-mail, used in a form of tags e.g. {{phone_number}}
                type: object
                additionalProperties: true
                example: >
                  {
                    "phone_number": "123456",
                    "favorite_animal": "honeybadger"
                  }
...
  parameters:
...
    id:
      name: id
      in: path
      description: ID of the object
      required: true
      schema:
        type: integer
        format: int64
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Message"
          example:
            message: "You are not authorized to perform this request."
    UnexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Message"
          example:
            message: "Something unexpected happened. Please try again."
  schemas:
    Message:
      type: object
      properties:
        message:
          type: string
          description: Response message.
    ValidationErrors:
      allOf:
        - $ref: "#/components/schemas/Message"
        - type: object
          description: Validation errors
          additionalProperties:
            type: string
...
    BaseModel:
      type: object
      properties:
        id:
          description: The ID of the resource.
          type: integer
          format: int64
          example: 111
        created_at:
          description: The date and time when the resource was created.
          type: string
          format: date-time
        updated_at:
          description: The date and time when the resource was last updated.
          type: string
          format: date-time
...
    BaseSubscriber:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            name:
              description: The name of the subscriber.
              type: string
              example: John Doe
            email:
              description: The email of the subscriber.
              type: string
              example: john@example.com
            metadata:
              description: Metadata associated with the subscriber.
              type: object
              additionalProperties:
                type: string
              example: >
                {
                  "phone_number": "123456789",
                  "favorite_animal": "honeybadger"
                }
            blacklisted:
              description: Flag that indicates if the subscriber has been blacklisted (unsubscribed, marked email as spam, bounced).
              type: boolean
            active:
              description: Flag that indicates if the subscriber is active or not (we send e-mails only to active subscribers).
              type: boolean
    Subscriber:
      allOf:
        - $ref: "#/components/schemas/BaseSubscriber"
        - type: object
          properties:
            segments:
              type: array
              items:
                $ref: "#/components/schemas/Group"
    BaseGroup:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          properties:
            name:
              description: The name of the group.
              type: string
              example: Foo Group
    Group:
      allOf:
        - $ref: "#/components/schemas/BaseGroup"
        - type: object
          properties:
            subscribers_in_segment:
              description: Number of subscribers inside the group.
              type: integer
              format: int64
              example: 123
...
  securitySchemes:
    api_key:
      type: apiKey
      name: X-API-Key
      in: header